APPDEV-1166 Package frinends to renamed to friends.
