+++ app/docs/nu/yona/app/ui/friends/TimelineFragment.html
@@ -3,7 +3,7 @@
 <html lang="en">
 <head>
 <!-- Generated by javadoc (1.8.0_25) on Sun May 01 13:07:48 IST 2016 -->
-<title>AddFriendManually</title>
+<title>TimelineFragment</title>
 <meta name="date" content="2016-05-01">
 <link rel="stylesheet" type="text/css" href="../../../../../stylesheet.css" title="Style">
 <script type="text/javascript" src="../../../../../script.js"></script>
@@ -12,7 +12,7 @@
 <script type="text/javascript"><!--
     try {
         if (location.href.indexOf('is-external=true') == -1) {
-            parent.document.title="AddFriendManually";
+            parent.document.title="TimelineFragment";
         }
     }
     catch(err) {
@@ -48,12 +48,12 @@ var activeTableTab = "activeTableTab";
 </div>
 <div class="subNav">
 <ul class="navList">
-<li><a href="../../../../../nu/yona/app/ui/frinends/AddFriendFragment.html" title="class in nu.yona.app.ui.frinends"><span class="typeNameLink">Prev&nbsp;Class</span></a></li>
-<li><a href="../../../../../nu/yona/app/ui/frinends/FriendsFragment.html" title="class in nu.yona.app.ui.frinends"><span class="typeNameLink">Next&nbsp;Class</span></a></li>
+<li><a href="../../../../../nu/yona/app/ui/frinends/OverviewFragment.html" title="class in nu.yona.app.ui.friends"><span class="typeNameLink">Prev&nbsp;Class</span></a></li>
+<li>Next&nbsp;Class</li>
 </ul>
 <ul class="navList">
-<li><a href="../../../../../index.html?nu/yona/app/ui/frinends/AddFriendManually.html" target="_top">Frames</a></li>
-<li><a href="AddFriendManually.html" target="_top">No&nbsp;Frames</a></li>
+<li><a href="../../../../../index.html?nu/yona/app/ui/frinends/TimelineFragment.html" target="_top">Frames</a></li>
+<li><a href="TimelineFragment.html" target="_top">No&nbsp;Frames</a></li>
 </ul>
 <ul class="navList" id="allclasses_navbar_top">
 <li><a href="../../../../../allclasses-noframe.html">All&nbsp;Classes</a></li>
@@ -91,8 +91,8 @@ var activeTableTab = "activeTableTab";
 <!-- ========= END OF TOP NAVBAR ========= -->
 <!-- ======== START OF CLASS DATA ======== -->
 <div class="header">
-<div class="subTitle">nu.yona.app.ui.frinends</div>
-<h2 title="Class AddFriendManually" class="title">Class AddFriendManually</h2>
+<div class="subTitle">nu.yona.app.ui.friends</div>
+<h2 title="Class TimelineFragment" class="title">Class TimelineFragment</h2>
 </div>
 <div class="contentContainer">
 <ul class="inheritance">
@@ -102,7 +102,7 @@ var activeTableTab = "activeTableTab";
 <li><a href="../../../../../nu/yona/app/ui/BaseFragment.html" title="class in nu.yona.app.ui">nu.yona.app.ui.BaseFragment</a></li>
 <li>
 <ul class="inheritance">
-<li>nu.yona.app.ui.frinends.AddFriendManuallyFragment</li>
+<li>nu.yona.app.ui.friends.TimelineFragment</li>
 </ul>
 </li>
 </ul>
@@ -113,9 +113,9 @@ var activeTableTab = "activeTableTab";
 <li class="blockList">
 <hr>
 <br>
-<pre>public class <span class="typeNameLabel">AddFriendManually</span>
+<pre>public class <span class="typeNameLabel">TimelineFragment</span>
 extends <a href="../../../../../nu/yona/app/ui/BaseFragment.html" title="class in nu.yona.app.ui">BaseFragment</a></pre>
-<div class="block">Created by kinnarvasa on 27/04/16.</div>
+<div class="block">Created by kinnarvasa on 21/03/16.</div>
 </li>
 </ul>
 </div>
@@ -134,7 +134,7 @@ extends <a href="../../../../../nu/yona/app/ui/BaseFragment.html" title="class i
 <th class="colOne" scope="col">Constructor and Description</th>
 </tr>
 <tr class="altColor">
-<td class="colOne"><code><span class="memberNameLink"><a href="../../../../../nu/yona/app/ui/frinends/AddFriendManually.html#AddFriendManually--">AddFriendManually</a></span>()</code>&nbsp;</td>
+<td class="colOne"><code><span class="memberNameLink"><a href="../../../../../nu/yona/app/ui/frinends/TimelineFragment.html#TimelineFragment--">TimelineFragment</a></span>()</code>&nbsp;</td>
 </tr>
 </table>
 </li>
@@ -153,7 +153,7 @@ extends <a href="../../../../../nu/yona/app/ui/BaseFragment.html" title="class i
 </tr>
 <tr id="i0" class="altColor">
 <td class="colFirst"><code>View</code></td>
-<td class="colLast"><code><span class="memberNameLink"><a href="../../../../../nu/yona/app/ui/frinends/AddFriendManually.html#onCreateView-LayoutInflater-ViewGroup-Bundle-">onCreateView</a></span>(LayoutInflater&nbsp;inflater,
+<td class="colLast"><code><span class="memberNameLink"><a href="../../../../../nu/yona/app/ui/frinends/TimelineFragment.html#onCreateView-LayoutInflater-ViewGroup-Bundle-">onCreateView</a></span>(LayoutInflater&nbsp;inflater,
             ViewGroup&nbsp;container,
             Bundle&nbsp;savedInstanceState)</code>&nbsp;</td>
 </tr>
@@ -179,13 +179,13 @@ extends <a href="../../../../../nu/yona/app/ui/BaseFragment.html" title="class i
 <!--   -->
 </a>
 <h3>Constructor Detail</h3>
-<a name="AddFriendManually--">
+<a name="TimelineFragment--">
 <!--   -->
 </a>
 <ul class="blockListLast">
 <li class="blockList">
-<h4>AddFriendManually</h4>
-<pre>public&nbsp;AddFriendManually()</pre>
+<h4>TimelineFragment</h4>
+<pre>public&nbsp;TimelineFragment()</pre>
 </li>
 </ul>
 </li>
@@ -237,12 +237,12 @@ public&nbsp;View&nbsp;onCreateView(LayoutInflater&nbsp;inflater,
 </div>
 <div class="subNav">
 <ul class="navList">
-<li><a href="../../../../../nu/yona/app/ui/frinends/AddFriendFragment.html" title="class in nu.yona.app.ui.frinends"><span class="typeNameLink">Prev&nbsp;Class</span></a></li>
-<li><a href="../../../../../nu/yona/app/ui/frinends/FriendsFragment.html" title="class in nu.yona.app.ui.frinends"><span class="typeNameLink">Next&nbsp;Class</span></a></li>
+<li><a href="../../../../../nu/yona/app/ui/frinends/OverviewFragment.html" title="class in nu.yona.app.ui.friends"><span class="typeNameLink">Prev&nbsp;Class</span></a></li>
+<li>Next&nbsp;Class</li>
 </ul>
 <ul class="navList">
-<li><a href="../../../../../index.html?nu/yona/app/ui/frinends/AddFriendManually.html" target="_top">Frames</a></li>
-<li><a href="AddFriendManually.html" target="_top">No&nbsp;Frames</a></li>
+<li><a href="../../../../../index.html?nu/yona/app/ui/frinends/TimelineFragment.html" target="_top">Frames</a></li>
+<li><a href="TimelineFragment.html" target="_top">No&nbsp;Frames</a></li>
 </ul>
 <ul class="navList" id="allclasses_navbar_bottom">
 <li><a href="../../../../../allclasses-noframe.html">All&nbsp;Classes</a></li>
