+++ app/docs/nu/yona/app/ui/friends/package-frame.html
@@ -1,0 +1,24 @@
+<!DOCTYPE HTML PUBLIC "-//W3C//DTD HTML 4.01 Transitional//EN" "http://www.w3.org/TR/html4/loose.dtd">
+<!-- NewPage -->
+<html lang="en">
+<head>
+<!-- Generated by javadoc (1.8.0_25) on Sun May 01 13:07:49 IST 2016 -->
+<title>nu.yona.app.ui.friends</title>
+<meta name="date" content="2016-05-01">
+<link rel="stylesheet" type="text/css" href="../../../../../stylesheet.css" title="Style">
+<script type="text/javascript" src="../../../../../script.js"></script>
+</head>
+<body>
+<h1 class="bar"><a href="../../../../../nu/yona/app/ui/frinends/package-summary.html" target="classFrame">nu.yona.app.ui.friends</a></h1>
+<div class="indexContainer">
+<h2 title="Classes">Classes</h2>
+<ul title="Classes">
+<li><a href="AddFriendFragment.html" title="class in nu.yona.app.ui.friends" target="classFrame">AddFriendFragment</a></li>
+<li><a href="AddFriendManually.html" title="class in nu.yona.app.ui.friends" target="classFrame">AddFriendManually</a></li>
+<li><a href="FriendsFragment.html" title="class in nu.yona.app.ui.friends" target="classFrame">FriendsFragment</a></li>
+<li><a href="OverviewFragment.html" title="class in nu.yona.app.ui.friends" target="classFrame">OverviewFragment</a></li>
+<li><a href="TimelineFragment.html" title="class in nu.yona.app.ui.friends" target="classFrame">TimelineFragment</a></li>
+</ul>
+</div>
+</body>
+</html>
