+++ app/docs/overview-frame.html
@@ -1,47 +1,76 @@
-<!DOCTYPE HTML PUBLIC "-//W3C//DTD HTML 4.01 Transitional//EN" "http://www.w3.org/TR/html4/loose.dtd">
+<!DOCTYPE HTML PUBLIC "-//W3C//DTD HTML 4.01 Transitional//EN"
+        "http://www.w3.org/TR/html4/loose.dtd">
 <!-- NewPage -->
 <html lang="en">
 <head>
-<!-- Generated by javadoc (1.8.0_25) on Sun May 01 13:07:49 IST 2016 -->
-<title>Overview List</title>
-<meta name="date" content="2016-05-01">
-<link rel="stylesheet" type="text/css" href="stylesheet.css" title="Style">
-<script type="text/javascript" src="script.js"></script>
+    <!-- Generated by javadoc (1.8.0_25) on Sun May 01 13:07:49 IST 2016 -->
+    <title>Overview List</title>
+    <meta name="date" content="2016-05-01">
+    <link rel="stylesheet" type="text/css" href="stylesheet.css" title="Style">
+    <script type="text/javascript" src="script.js"></script>
 </head>
 <body>
-<div class="indexHeader"><span><a href="allclasses-frame.html" target="packageFrame">All&nbsp;Classes</a></span></div>
+<div class="indexHeader"><span><a href="allclasses-frame.html" target="packageFrame">All&nbsp;Classes</a></span>
+</div>
 <div class="indexContainer">
-<h2 title="Packages">Packages</h2>
-<ul title="Packages">
-<li><a href="nu/yona/app/package-frame.html" target="packageFrame">nu.yona.app</a></li>
-<li><a href="nu/yona/app/api/package-frame.html" target="packageFrame">nu.yona.app.api</a></li>
-<li><a href="nu/yona/app/api/db/package-frame.html" target="packageFrame">nu.yona.app.api.db</a></li>
-<li><a href="nu/yona/app/api/manager/package-frame.html" target="packageFrame">nu.yona.app.api.manager</a></li>
-<li><a href="nu/yona/app/api/manager/dao/package-frame.html" target="packageFrame">nu.yona.app.api.manager.dao</a></li>
-<li><a href="nu/yona/app/api/manager/impl/package-frame.html" target="packageFrame">nu.yona.app.api.manager.impl</a></li>
-<li><a href="nu/yona/app/api/manager/network/package-frame.html" target="packageFrame">nu.yona.app.api.manager.network</a></li>
-<li><a href="nu/yona/app/api/model/package-frame.html" target="packageFrame">nu.yona.app.api.model</a></li>
-<li><a href="nu/yona/app/api/receiver/package-frame.html" target="packageFrame">nu.yona.app.api.receiver</a></li>
-<li><a href="nu/yona/app/api/service/package-frame.html" target="packageFrame">nu.yona.app.api.service</a></li>
-<li><a href="nu/yona/app/api/utils/package-frame.html" target="packageFrame">nu.yona.app.api.utils</a></li>
-<li><a href="nu/yona/app/customview/package-frame.html" target="packageFrame">nu.yona.app.customview</a></li>
-<li><a href="nu/yona/app/enums/package-frame.html" target="packageFrame">nu.yona.app.enums</a></li>
-<li><a href="nu/yona/app/listener/package-frame.html" target="packageFrame">nu.yona.app.listener</a></li>
-<li><a href="nu/yona/app/state/package-frame.html" target="packageFrame">nu.yona.app.state</a></li>
-<li><a href="nu/yona/app/support/package-frame.html" target="packageFrame">nu.yona.app.support</a></li>
-<li><a href="nu/yona/app/ui/package-frame.html" target="packageFrame">nu.yona.app.ui</a></li>
-<li><a href="nu/yona/app/ui/challenges/package-frame.html" target="packageFrame">nu.yona.app.ui.challenges</a></li>
-<li><a href="nu/yona/app/ui/dashboard/package-frame.html" target="packageFrame">nu.yona.app.ui.dashboard</a></li>
-<li><a href="nu/yona/app/ui/frinends/package-frame.html" target="packageFrame">nu.yona.app.ui.frinends</a></li>
-<li><a href="nu/yona/app/ui/login/package-frame.html" target="packageFrame">nu.yona.app.ui.login</a></li>
-<li><a href="nu/yona/app/ui/message/package-frame.html" target="packageFrame">nu.yona.app.ui.message</a></li>
-<li><a href="nu/yona/app/ui/pincode/package-frame.html" target="packageFrame">nu.yona.app.ui.pincode</a></li>
-<li><a href="nu/yona/app/ui/profile/package-frame.html" target="packageFrame">nu.yona.app.ui.profile</a></li>
-<li><a href="nu/yona/app/ui/settings/package-frame.html" target="packageFrame">nu.yona.app.ui.settings</a></li>
-<li><a href="nu/yona/app/ui/signup/package-frame.html" target="packageFrame">nu.yona.app.ui.signup</a></li>
-<li><a href="nu/yona/app/ui/tour/package-frame.html" target="packageFrame">nu.yona.app.ui.tour</a></li>
-<li><a href="nu/yona/app/utils/package-frame.html" target="packageFrame">nu.yona.app.utils</a></li>
-</ul>
+    <h2 title="Packages">Packages</h2>
+    <ul title="Packages">
+        <li><a href="nu/yona/app/package-frame.html" target="packageFrame">nu.yona.app</a></li>
+        <li><a href="nu/yona/app/api/package-frame.html" target="packageFrame">nu.yona.app.api</a>
+        </li>
+        <li><a href="nu/yona/app/api/db/package-frame.html"
+               target="packageFrame">nu.yona.app.api.db</a></li>
+        <li><a href="nu/yona/app/api/manager/package-frame.html" target="packageFrame">nu.yona.app.api.manager</a>
+        </li>
+        <li><a href="nu/yona/app/api/manager/dao/package-frame.html" target="packageFrame">nu.yona.app.api.manager.dao</a>
+        </li>
+        <li><a href="nu/yona/app/api/manager/impl/package-frame.html" target="packageFrame">nu.yona.app.api.manager.impl</a>
+        </li>
+        <li><a href="nu/yona/app/api/manager/network/package-frame.html" target="packageFrame">nu.yona.app.api.manager.network</a>
+        </li>
+        <li><a href="nu/yona/app/api/model/package-frame.html" target="packageFrame">nu.yona.app.api.model</a>
+        </li>
+        <li><a href="nu/yona/app/api/receiver/package-frame.html" target="packageFrame">nu.yona.app.api.receiver</a>
+        </li>
+        <li><a href="nu/yona/app/api/service/package-frame.html" target="packageFrame">nu.yona.app.api.service</a>
+        </li>
+        <li><a href="nu/yona/app/api/utils/package-frame.html" target="packageFrame">nu.yona.app.api.utils</a>
+        </li>
+        <li><a href="nu/yona/app/customview/package-frame.html" target="packageFrame">nu.yona.app.customview</a>
+        </li>
+        <li><a href="nu/yona/app/enums/package-frame.html"
+               target="packageFrame">nu.yona.app.enums</a></li>
+        <li><a href="nu/yona/app/listener/package-frame.html" target="packageFrame">nu.yona.app.listener</a>
+        </li>
+        <li><a href="nu/yona/app/state/package-frame.html"
+               target="packageFrame">nu.yona.app.state</a></li>
+        <li><a href="nu/yona/app/support/package-frame.html" target="packageFrame">nu.yona.app.support</a>
+        </li>
+        <li><a href="nu/yona/app/ui/package-frame.html" target="packageFrame">nu.yona.app.ui</a>
+        </li>
+        <li><a href="nu/yona/app/ui/challenges/package-frame.html" target="packageFrame">nu.yona.app.ui.challenges</a>
+        </li>
+        <li><a href="nu/yona/app/ui/dashboard/package-frame.html" target="packageFrame">nu.yona.app.ui.dashboard</a>
+        </li>
+        <li><a href="nu/yona/app/ui/frinends/package-frame.html" target="packageFrame">nu.yona.app.ui.friends</a>
+        </li>
+        <li><a href="nu/yona/app/ui/login/package-frame.html" target="packageFrame">nu.yona.app.ui.login</a>
+        </li>
+        <li><a href="nu/yona/app/ui/message/package-frame.html" target="packageFrame">nu.yona.app.ui.message</a>
+        </li>
+        <li><a href="nu/yona/app/ui/pincode/package-frame.html" target="packageFrame">nu.yona.app.ui.pincode</a>
+        </li>
+        <li><a href="nu/yona/app/ui/profile/package-frame.html" target="packageFrame">nu.yona.app.ui.profile</a>
+        </li>
+        <li><a href="nu/yona/app/ui/settings/package-frame.html" target="packageFrame">nu.yona.app.ui.settings</a>
+        </li>
+        <li><a href="nu/yona/app/ui/signup/package-frame.html" target="packageFrame">nu.yona.app.ui.signup</a>
+        </li>
+        <li><a href="nu/yona/app/ui/tour/package-frame.html" target="packageFrame">nu.yona.app.ui.tour</a>
+        </li>
+        <li><a href="nu/yona/app/utils/package-frame.html"
+               target="packageFrame">nu.yona.app.utils</a></li>
+    </ul>
 </div>
 <p>&nbsp;</p>
 </body>
