+++ app/src/main/java/nu/yona/app/ui/BaseFragment.java
@@ -23,9 +23,9 @@ import nu.yona.app.customview.YonaFontTextView;
 import nu.yona.app.ui.challenges.ChallengesFragment;
 import nu.yona.app.ui.challenges.ChallengesGoalDetailFragment;
 import nu.yona.app.ui.dashboard.DashboardFragment;
-import nu.yona.app.ui.frinends.AddFriendFragment;
-import nu.yona.app.ui.frinends.FriendsFragment;
-import nu.yona.app.ui.frinends.FriendsRequestFragment;
+import nu.yona.app.ui.friends.AddFriendFragment;
+import nu.yona.app.ui.friends.FriendsFragment;
+import nu.yona.app.ui.friends.FriendsRequestFragment;
 import nu.yona.app.ui.message.NotificationFragment;
 import nu.yona.app.ui.settings.PrivacyFragment;
 import nu.yona.app.ui.settings.SettingsFragment;
