+++ app/src/main/java/nu/yona/app/ui/friends/OnFriendsItemClickListener.java
@@ -6,7 +6,7 @@
  * file, You can obtain one at https://mozilla.org/MPL/2.0/.
  */
 
-package nu.yona.app.ui.frinends;
+package nu.yona.app.ui.friends;
 
 import android.view.View;
